Dialog for changing proximity initial commit. TBD: favorite socials of each category + network management.
