+++ composeApp/src/commonMain/kotlin/ui/home/HomeRepository.kt
@@ -5,16 +5,12 @@ import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import data.io.base.BaseResponse
 import data.io.social.network.conversation.matrix.ConversationRoomIO
-import data.io.social.network.conversation.matrix.MatrixEventContent
-import data.io.social.network.conversation.matrix.RoomNotificationsCount
-import data.io.social.network.conversation.matrix.RoomSummary
 import data.io.social.network.conversation.matrix.RoomType
 import data.io.social.network.conversation.matrix.RoomsResponseIO
 import data.io.social.network.request.NetworkListResponse
 import data.io.user.matrix.SyncResponse
 import database.dao.ConversationRoomDao
 import database.dao.MatrixPagingMetaDao
-import database.dao.NetworkItemDao
 import dev.gitlive.firebase.Firebase
 import dev.gitlive.firebase.auth.auth
 import io.ktor.client.HttpClient
@@ -36,7 +32,6 @@ import ui.network.list.NetworkListRepository.Companion.proximityDemoData
 class HomeRepository(
     private val httpClient: HttpClient,
     private val conversationRoomDao: ConversationRoomDao,
-    private val networkDao: NetworkItemDao,
     private val pagingMetaDao: MatrixPagingMetaDao
 ) {
     /** returns a list of network list */
@@ -60,7 +55,7 @@ class HomeRepository(
                         }
                     }
                 )
-            }.takeIf { it.success?.data != null } ?: BaseResponse.Success(DEMO_SYNC)
+            }
         }
     }
 
@@ -125,18 +120,17 @@ class HomeRepository(
         )
     }
 
-    /** Updates a network connection */
-    suspend fun patchNetworkConnection(publicId: String, proximity: Float): BaseResponse<Any> {
+    /** Updates a conversation's proximity */
+    suspend fun patchConversationProximity(id: String, proximity: Float): BaseResponse<Any> {
         return withContext(Dispatchers.IO) {
-            networkDao.updateProximity(
-                ownerPublicId = Firebase.auth.currentUser?.uid,
-                publicId = publicId,
+            conversationRoomDao.updateProximity(
+                id = id,
                 proximity = proximity
             )
 
             httpClient.safeRequest<NetworkListResponse> {
                 patch(
-                    urlString = "/api/v1/social/network/users/{$publicId}",
+                    urlString = "/api/v1/social/conversation/$id",
                     block = {
                         setBody(SocialConnectionUpdate(proximity = proximity))
                     }
@@ -147,36 +141,5 @@ class HomeRepository(
 
     companion object {
         const val INITIAL_BATCH = "initial_batch"
-
-        private val DEMO_SYNC = SyncResponse(
-            nextBatch = null,
-            rooms = RoomsResponseIO(
-                join = hashMapOf(
-                    "1" to ConversationRoomIO(
-                        id = "1",
-                        summary = RoomSummary(
-                            heroes = listOf("1"),
-                            lastMessage = MatrixEventContent.RoomMessageEvent(body = "Hey, what's up?"),
-                            joinedMemberCount = 2
-                        ),
-                        proximity = 5f
-                    ),
-                    "2" to ConversationRoomIO(
-                        id = "2",
-                        unreadNotifications = RoomNotificationsCount(highlightCount = 2),
-                        summary = RoomSummary(
-                            heroes = listOf("2"),
-                            canonicalAlias = "Gamer's room",
-                            lastMessage = MatrixEventContent.RoomMessageEvent(body = "That's terrible:D"),
-                            joinedMemberCount = 2
-                        ),
-                        proximity = 2f
-                    )
-                ),
-                invite = hashMapOf(),
-                knock = hashMapOf(),
-                leave = hashMapOf()
-            )
-        )
     }
 }
